+++ crates/ui/lib/widgets/transport/transport_controls.dart
@@ -1,19 +1,14 @@
-import 'dart:developer';
-
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:mizer/api/contracts/transport.dart';
-import 'package:mizer/api/contracts/ui.dart';
-import 'package:mizer/extensions/context_state_extensions.dart';
 import 'package:mizer/protos/transport.pb.dart';
-import 'package:mizer/widgets/controls/select.dart';
 import 'package:mizer/widgets/hoverable.dart';
 import 'package:mizer/widgets/panel.dart';
-import 'package:mizer/widgets/text_field_focus.dart';
 import 'package:mizer/widgets/transport/beat_indicator.dart';
 
+import 'command_line.dart';
+import 'fps_control.dart';
+import 'speed_control.dart';
 import 'time_control.dart';
 
 const double TRANSPORT_CONTROLS_HEIGHT = 40;
@@ -115,177 +110,6 @@ class _TransportControlsState extends State<TransportControls> {
   }
 }
 
-class FpsSelector extends StatelessWidget {
-  final Stream<Transport> transportStream;
-
-  const FpsSelector({super.key, required this.transportStream});
-
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<double>(
-        stream: transportStream.map((event) => event.fps).distinct(),
-        builder: (context, snapshot) {
-          return Container(
-            width: 96,
-            height: TRANSPORT_CONTROLS_HEIGHT,
-            padding: const EdgeInsets.all(4),
-            child: MizerSelect<double>(
-              openTop: true,
-              options: [
-                SelectOption(label: "30 FPS", value: 30),
-                SelectOption(label: "60 FPS", value: 60),
-              ],
-              onChanged: (fps) => context.read<TransportApi>().setFPS(fps),
-              value: snapshot.data ?? 60,
-            ),
-          );
-        });
-  }
-}
-
-class SpeedControl extends StatelessWidget {
-  final Stream<double> bpm;
-
-  const SpeedControl(this.bpm, {Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-        width: 96,
-        child: StreamBuilder<double>(
-            stream: bpm,
-            initialData: 0,
-            builder: (context, snapshot) => Center(
-                  child: SpeedEditor(
-                      value: snapshot.data!,
-                      onUpdate: (bpm) => context.read<TransportApi>().setBPM(bpm)),
-                )));
-  }
-}
-
-class SpeedEditor extends StatefulWidget {
-  final double value;
-  final Function(double) onUpdate;
-
-  static final floatsOnly = FilteringTextInputFormatter.allow(RegExp(r'[0-9.]'));
-
-  const SpeedEditor({required this.value, required this.onUpdate, Key? key}) : super(key: key);
-
-  @override
-  State<SpeedEditor> createState() => _SpeedEditorState(value);
-}
-
-class _SpeedEditorState extends State<SpeedEditor> {
-  final FocusNode focusNode = FocusNode(debugLabel: "SpeedEditor");
-  final TextEditingController controller = TextEditingController();
-  double value;
-  bool isEditing = false;
-
-  _SpeedEditorState(this.value) {
-    this.controller.text = value.toString();
-  }
-
-  @override
-  void didUpdateWidget(SpeedEditor oldWidget) {
-    if (oldWidget.value != widget.value && widget.value != this.value) {
-      this.controller.text = widget.value.toString();
-      setState(() {
-        value = widget.value;
-      });
-    }
-    super.didUpdateWidget(oldWidget);
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _setValue(value);
-    this.focusNode.addListener(() {
-      if (!focusNode.hasFocus) {
-        var value = double.parse(controller.text);
-        this._setValue(value);
-        setState(() {
-          this.isEditing = false;
-        });
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return this.isEditing ? _editView(context) : _readView(context);
-  }
-
-  @override
-  void dispose() {
-    focusNode.dispose();
-    controller.dispose();
-    super.dispose();
-  }
-
-  Widget _readView(BuildContext context) {
-    var style = Theme.of(context).textTheme;
-    return MouseRegion(
-      cursor: SystemMouseCursors.resizeLeftRight,
-      child: Listener(
-        onPointerSignal: (e) {
-          if (e is PointerScrollEvent) {
-            var delta = e.scrollDelta.dy > 0 ? -1 : 1;
-            var next = this.value + delta;
-            _dragValue(double.parse(next.toStringAsFixed(3)));
-          }
-        },
-        child: GestureDetector(
-          onHorizontalDragUpdate: (update) {
-            var delta = update.primaryDelta ?? 0;
-            var next = this.value + delta;
-            _dragValue(double.parse(next.toStringAsFixed(3)));
-          },
-          onTap: () => setState(() => this.isEditing = true),
-          child: Text(value.toStringAsFixed(2),
-              textAlign: TextAlign.center, style: style.headlineSmall!.copyWith(fontFamily: "RobotoMono", fontSize: 20)),
-        ),
-      ),
-    );
-  }
-
-  Widget _editView(BuildContext context) {
-    TextStyle textStyle = Theme.of(context).textTheme.headlineSmall!;
-
-    return EditableText(
-      focusNode: focusNode,
-      controller: controller,
-      cursorColor: Colors.black87,
-      backgroundCursorColor: Colors.black12,
-      style: textStyle,
-      textAlign: TextAlign.center,
-      selectionColor: Colors.black38,
-      keyboardType: TextInputType.number,
-      autofocus: true,
-      inputFormatters: [
-        SpeedEditor.floatsOnly,
-        FilteringTextInputFormatter.singleLineFormatter,
-      ],
-    );
-  }
-
-  void _dragValue(double value) {
-    this.controller.text = this.value.toString();
-    _setValue(value);
-  }
-
-  void _setValue(double value) {
-    log("setValue $value", name: "SpeedEditor");
-    value = value.clamp(0, 300);
-    setState(() {
-      this.value = value;
-    });
-    if (widget.value != value) {
-      widget.onUpdate(value);
-    }
-  }
-}
-
 class TransportControl extends StatelessWidget {
   final Stream<TransportState> state;
 
@@ -343,84 +167,5 @@ class _Divider extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(width: 2, color: Colors.black26);
-  }
-}
-
-class CommandLineInput extends StatefulWidget {
-  const CommandLineInput({super.key});
-
-  @override
-  State<CommandLineInput> createState() => _CommandLineInputState();
-}
-
-class _CommandLineInputState extends State<CommandLineInput> {
-  TextEditingController controller = TextEditingController(text: '');
-  FocusNode focusNode = FocusNode();
-  bool success = false;
-  bool error = false;
-
-  @override
-  Widget build(BuildContext context) {
-    TextTheme textTheme = Theme.of(context).textTheme;
-    return Container(
-      height: TRANSPORT_CONTROLS_HEIGHT - 8,
-      decoration: BoxDecoration(
-        color: Colors.white10,
-        borderRadius: BorderRadius.circular(2),
-        border: Border.all(color: Colors.black26),
-      ),
-      margin: const EdgeInsets.all(4),
-      padding: const EdgeInsets.all(4),
-      child: Stack(
-        children: [
-          if (!focusNode.hasFocus && controller.text.isEmpty)
-            Center(
-                child: Text("Commandline",
-                    style: textTheme.bodyMedium!.copyWith(color: Colors.white54))),
-          if (success) Positioned(child: Icon(Icons.check, color: Colors.green), right: 4),
-          if (error)
-            Positioned(child: Icon(Icons.warning_amber_outlined, color: Colors.red), right: 4),
-          Center(
-              child: TextFieldFocus(
-            child: EditableText(
-              controller: controller,
-              focusNode: focusNode,
-              style: textTheme.bodyMedium!,
-              autocorrect: false,
-              showCursor: true,
-              cursorColor: Colors.black87,
-              backgroundCursorColor: Colors.black12,
-              textAlign: TextAlign.start,
-              selectionColor: Colors.black38,
-              keyboardType: TextInputType.text,
-              inputFormatters: [FilteringTextInputFormatter.singleLineFormatter],
-              onSubmitted: (value) {
-                execute(value);
-                controller.clear();
-                focusNode.requestFocus();
-              },
-            ),
-          )),
-        ],
-      ),
-    );
-  }
-
-  void execute(String cmd) {
-    context.read<UiApi>().commandLineExecute(cmd).then((value) {
-      setIconState(success: true);
-      context.refreshAllStates();
-      Future.delayed(Duration(seconds: 3)).then((value) => setIconState());
-    }).catchError((err) {
-      setIconState(error: true);
-      Future.delayed(Duration(seconds: 3)).then((value) => setIconState());
-    });
-  }
-
-  void setIconState({ bool success = false, bool error = false }) {
-    setState(() {
-      this.success = success;
-      this.error = error;
-    });
   }
 }
